+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.10.0
 
 - [BC] The `getDefinitionByType` now throws an exception if the definition does not exist, instead of returning null.
+- Added two new `findByIdOrFail` and `findByHandleOrFail` on the metaobject repository.
 
 ## 0.9.0